// generated/src/app/api-models/inst-patient-liab-edi-w.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class InstPatientLiabEdiW {
  seqPrediId: number;
  seqClaimId: number;
  lineNumber: number;
  subLineCode: string;
  ruleType: string;
  ruleId: string;
  liabilityAmt: number;
  liabilityReason: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}